Moved Aliot.notif and Aliot.notif_err to the builtin module
